\name{function_hours}
\alias{function_hours}
\title{
Hours
}
\description{
It aims to convert minutes into hours
}
\usage{
function_hours(minutes)
}

\arguments{
  \item{minutes}{
number of minutes
}
}
\details{
Convert minutes into hours 
}


\examples{
minutes<-560
function_hours(minutes)
# return de value "09:20"
}
\value{
Returns a number corresponding to the number of hours. 
}

\author{
Josefina Urquiza, Maria Florencia Tames, Maela Lupo, Alfredo Rigalli

Maintainer: Josefina Urquiza <jurquiza@conicet.gov.ar>
}

\keyword{manip}
\concept{air}
\concept{pollution}
